refactor(cli): extract native TUI subcommands into cli_native.py
Phase 0 of making native harnesses pluggable: carve the 11 native
coding-agent subcommands (claude, codex, opencode, pi, cursor, kiro,
goose, hermes, antigravity, qwen, kimi) out of cli.py into a dedicated
cli_native.py so the follow-up registry-driven seam lands in a small,
focused module instead of a 14k-line file. Behavior-preserving.

- New omnigent/cli_common.py holds the decorator-time constants
  (RESUME_PICKER_SENTINEL, CLAUDE_STARTUP_PROFILE_ENV_VAR) and
  reject_native_on_windows. It is a leaf module (imports nothing from
  omnigent.cli), so both cli.py and cli_native.py can import it without a
  cycle — required because Click evaluates command decorators at import
  time.
- omnigent/cli_native.py exposes register_native_commands(cli), which
  cli.py calls at module bottom (after the group and shared launch
  helpers exist). Command bodies reach shared cli.py helpers through thin
  call-time proxies on the omnigent.cli module, which keeps this module
  free of a top-level omnigent.cli import (no cycle) and lets tests that
  monkeypatch omnigent.cli.<helper> still take effect.
- polly/debby (bundled example agents, not native TUIs) stay in cli.py,
  along with the shared helpers they and the native commands use.

Also drafts designs/harness-modular-registry-proposal.md (the doc the
harness_plugins.py comment already references), which lays out the full
NativeHarnessProvider plan and the phasing this commit begins.

Test plan: tests/cli/test_cli.py (244), test_chat.py/test_import.py/
test_runner_startup.py (137) all pass; ruff format+check and the
pre-commit file hooks pass; `omnigent <tool> --help` renders for all 11.
